+++ src/main/java/br/com/laercioskt/backend/service/UserService.java
@@ -60,7 +60,7 @@ public class UserService {
 
     private Category createCategory(int i) {
         Category category = new Category();
-        category.setId(i);
+        category.setId((long) i);
         category.setName("Name " + i);
         return categoryRepository.save(category);
     }
